requirements.txt
fastapi
uvicorn[standard]
openai
pandas
duckdb
httpx
beautifulsoup4
matplotlib
python-dotenv
